+++ greenfield/eldenring/data.py
@@ -5538,4 +5538,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:47e3bd802664766e1f113855e071d78932d7aad3379f8be3a6ab0fd1fb572e5e', 'module': 'data.py', 'body_sha256': 'sha256:2d24587b24f792f21ec7bf6629a3192e4c14383a1e2766cc1630cc14ecb86ce2'}
+_GEN_STAMP = {'inputs_hash': 'sha256:674eec30afe5b6c6a9b47466d035fe68fb19c99c6aaec77a725ecd3dce067e35', 'module': 'data.py', 'body_sha256': 'sha256:2d24587b24f792f21ec7bf6629a3192e4c14383a1e2766cc1630cc14ecb86ce2'}